Revert "fix: swapper transition states (#7917)"
This reverts commit 4dffea961115bdcade62b3b2c902136d6b923d1a.
