+++ src/components/MultiHopTrade/components/SharedTradeInput/SharedTradeInputFooter.tsx
@@ -16,10 +16,9 @@ import { ReceiveSummary } from 'components/MultiHopTrade/components/TradeInput/c
 import { RecipientAddress } from 'components/MultiHopTrade/components/TradeInput/components/RecipientAddress'
 import { WithLazyMount } from 'components/MultiHopTrade/components/TradeInput/components/WithLazyMount'
 import { Text } from 'components/Text'
-import { useAccountsFetchQuery } from 'context/AppProvider/hooks/useAccountsFetchQuery'
 import { useWallet } from 'hooks/useWallet/useWallet'
 import { useWalletSupportsChain } from 'hooks/useWalletSupportsChain/useWalletSupportsChain'
-import { selectFeeAssetById } from 'state/slices/selectors'
+import { selectFeeAssetById, selectIsAccountsMetadataLoading } from 'state/slices/selectors'
 import { useAppSelector } from 'state/store'
 import { breakpoints } from 'theme/theme'
 
@@ -92,8 +91,7 @@ export const SharedTradeInputFooter = ({
   const buyAssetFeeAsset = useAppSelector(state =>
     selectFeeAssetById(state, buyAsset?.assetId ?? ''),
   )
-
-  const { isFetching: isAccountsMetadataLoading } = useAccountsFetchQuery()
+  const isAccountsMetadataLoading = useAppSelector(selectIsAccountsMetadataLoading)
   const walletSupportsBuyAssetChain = useWalletSupportsChain(buyAsset.chainId, wallet)
 
   const displayManualAddressEntry = useMemo(() => {
@@ -197,6 +195,7 @@ export const SharedTradeInputFooter = ({
           shouldForceManualAddressEntry={shouldForceManualAddressEntry}
           component={ManualAddressEntry}
           description={manualAddressEntryDescription}
+          chainId={buyAsset.chainId}
         />
 
         <ButtonWalletPredicate
